[Chore] move `panic` to a new file `panic.ts`
