+++ test/utils/helpers.ts
@@ -2,7 +2,8 @@ import * as assert from 'assert';
 import { DataBuffer } from 'src/base/common/files/buffer';
 import { FileType, IResolvedFileStat } from 'src/base/common/files/file';
 import { URI } from 'src/base/common/files/uri';
-import { AsyncResult, panic } from "src/base/common/result";
+import { AsyncResult } from "src/base/common/result";
+import { panic } from "src/base/common/utilities/panic";
 import { repeat } from "src/base/common/utilities/async";
 import { Random } from "src/base/common/utilities/random";
 import { NestedArray, TreeLike } from "src/base/common/utilities/type";
